rename _base_rules to _base_verified
